Level fix

+++ Modules/SoulGuard/Watchdogs/PowerWatchdog.cpp
@@ -18,7 +18,7 @@ void PowerWatchdog::check()
 		return;
 	}
 
-	uint16_t voltage = get_system_power();
+	uint32_t voltage = get_system_power();
 
 	if (STM_MIN_VOLTAGEx10 <= voltage && voltage <= STM_MAX_VOLTAGEx10) {
 		reset_error(POWER_ERROR);
